requirements.txt
streamlit
pandas
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests